Add due date filtering (allow-inactive) Revamp test approach (looks better? 😬)
